Make unsignedBigInteger For sorting_classification_id In Posts Migration

+++ database/migrations/2023_08_12_101422_create_posts_table.php
@@ -63,7 +63,7 @@ return new class extends Migration
             $table->tinyInteger('sorted')->default(0);
             $table->integer('sorter')->nullable();
             $table->string('sorted_date')->nullable();
-            $table->integer('sorting_classification_id')->nullable();
+            $table->unsignedBigInteger('sorting_classification_id')->nullable();
             $table->foreign('sorting_classification_id')->references('id')->on('users');
 
             $table->timestamps();
